fix: correct plugin build path in Makefile
The plugins were being built in plugins/build/plugins/ but needed to be
copied to build/plugins/ for the application to find them. Updated the
plugins target to copy the built plugins to the correct location.

This fixes the 'unsupported workload type' error when trying to use
plugin-based workloads like ecommerce_mixed.